// g6-realestate-frontend/src/app/interfaces/vivienda-filtro.interface.ts
export interface ViviendaFiltroResponse{
    
    id: number;
    titulo: string;
    tipo: string;
    direccion: string;
    provincia: string;
    poblacion: string;
    avatar: string;
    precio: number;
    metrosCuadrados: number;
    numeroBanos: number;
    numeroHabitaciones: number;
    
}